0096.py
# Project 96. Object Detection in Images - MODERNIZED VERSION
# Description:
# This is a comprehensive object detection system supporting multiple state-of-the-art models
# including YOLOv8, YOLOv9, SSD MobileNet, and EfficientDet. Features include:
# - Modern web UI with Streamlit
# - Database storage and analytics
# - Batch processing capabilities
# - Performance metrics and evaluation
# - Multiple model support with easy switching

# 🚀 QUICK START:
# 1. Install dependencies: pip install -r requirements.txt
# 2. Run web interface: streamlit run app.py
# 3. Or use programmatically: python object_detector.py

# Import the modern object detection system
from object_detector import ObjectDetector
from database import DetectionDatabase, MockDataGenerator
import cv2
import numpy as np
from PIL import Image
import matplotlib.pyplot as plt

def demo_basic_detection():
    """Demonstrate basic object detection functionality"""
    print("🔍 Initializing Modern Object Detection System...")
    
    # Initialize detector with modern configuration
    detector = ObjectDetector("config.yaml")
    
    # Load YOLOv8 model (latest and fastest)
    if detector.load_model('yolo'):
        print("✅ YOLOv8 model loaded successfully!")
        
        # Create a sample image for demonstration
        # In practice, you would load your own image
        sample_image = create_sample_image()
        
        try:
            # Run detection
            print("🔍 Running object detection...")
            result = detector.detect_objects(sample_image)
            
            print(f"✅ Detection completed!")
            print(f"📊 Found {len(result['detections'])} objects")
            print(f"⚡ Processing time: {result.get('processing_time', 'N/A')} ms")
            print(f"🤖 Model used: {result['model']}")
            
            # Display detection results
            for i, detection in enumerate(result['detections']):
                print(f"  {i+1}. {detection['class_name']} (confidence: {detection['confidence']:.3f})")
            
            # Draw and save results
            output_image = detector.draw_detections(
                result['image'], 
                result['detections']
            )
            
            # Save result
            cv2.imwrite('detection_result.jpg', cv2.cvtColor(output_image, cv2.COLOR_RGB2BGR))
            print("💾 Result saved as 'detection_result.jpg'")
            
        except Exception as e:
            print(f"❌ Detection failed: {e}")
    
    # Also demonstrate SSD MobileNet
    if detector.load_model('ssd_mobilenet'):
        print("✅ SSD MobileNet model loaded successfully!")
    
    return detector

def create_sample_image():
    """Create a sample image for demonstration"""
    # Create a simple test image with some geometric shapes
    img = np.zeros((480, 640, 3), dtype=np.uint8)
    
    # Add some colored rectangles (simulating objects)
    cv2.rectangle(img, (100, 100), (200, 200), (0, 255, 0), -1)  # Green rectangle
    cv2.rectangle(img, (300, 150), (450, 300), (255, 0, 0), -1)  # Blue rectangle
    cv2.rectangle(img, (50, 300), (150, 450), (0, 0, 255), -1)   # Red rectangle
    
    # Add some text
    cv2.putText(img, "Sample Image", (250, 50), cv2.FONT_HERSHEY_SIMPLEX, 1, (255, 255, 255), 2)
    
    return img

def demo_database_integration():
    """Demonstrate database integration"""
    print("\n🗄️ Demonstrating Database Integration...")
    
    # Initialize database
    db = DetectionDatabase("demo_detection.db")
    
    # Generate sample data
    generator = MockDataGenerator(db)
    print("📊 Generating sample detection data...")
    
    image_ids = generator.generate_sample_images(5)
    generator.generate_sample_detections(image_ids)
    
    # Get statistics
    stats = db.get_detection_statistics()
    print(f"📈 Database Statistics:")
    print(f"   Total Images: {stats['total_images']}")
    print(f"   Total Detections: {stats['total_detections']}")
    print(f"   Model Statistics: {stats['model_statistics']}")
    
    # Search example
    results = db.search_detections(class_name='person', min_confidence=0.5, limit=5)
    print(f"🔍 Found {len(results)} person detections with confidence >= 0.5")
    
    db.close()
    print("✅ Database demo completed!")

def demo_batch_processing():
    """Demonstrate batch processing capabilities"""
    print("\n📦 Demonstrating Batch Processing...")
    
    detector = ObjectDetector()
    detector.load_model('yolo')
    
    # Create multiple sample images
    image_paths = []
    for i in range(3):
        img = create_sample_image()
        path = f"batch_image_{i}.jpg"
        cv2.imwrite(path, cv2.cvtColor(img, cv2.COLOR_RGB2BGR))
        image_paths.append(path)
    
    # Process batch
    print(f"🔄 Processing {len(image_paths)} images...")
    results = detector.batch_detect(image_paths)
    
    for i, result in enumerate(results):
        if 'error' not in result:
            print(f"  Image {i+1}: {len(result['detections'])} objects detected")
        else:
            print(f"  Image {i+1}: Error - {result['error']}")
    
    # Cleanup
    for path in image_paths:
        import os
        os.remove(path)
    
    print("✅ Batch processing demo completed!")

def main():
    """Main demonstration function"""
    print("=" * 60)
    print("🔍 MODERN OBJECT DETECTION SYSTEM DEMO")
    print("=" * 60)
    
    try:
        # Basic detection demo
        detector = demo_basic_detection()
        
        # Database integration demo
        demo_database_integration()
        
        # Batch processing demo
        demo_batch_processing()
        
        print("\n" + "=" * 60)
        print("🎉 ALL DEMOS COMPLETED SUCCESSFULLY!")
        print("=" * 60)
        print("\n🚀 NEXT STEPS:")
        print("1. Run 'streamlit run app.py' for the web interface")
        print("2. Upload your own images for detection")
        print("3. Explore the analytics dashboard")
        print("4. Check the database for stored results")
        print("\n📚 For more advanced usage, see:")
        print("   - object_detector.py: Core detection functionality")
        print("   - database.py: Database operations")
        print("   - app.py: Web interface")
        print("   - config.yaml: Configuration settings")
        
    except Exception as e:
        print(f"❌ Demo failed: {e}")
        print("💡 Make sure all dependencies are installed: pip install -r requirements.txt")

if __name__ == "__main__":
    main()

# 🧠 What This Modernized Project Demonstrates:
# ✅ Latest AI models: YOLOv8, YOLOv9, SSD MobileNet, EfficientDet
# ✅ Modern web interface with Streamlit
# ✅ Database storage and analytics
# ✅ Batch processing capabilities
# ✅ Performance metrics and evaluation
# ✅ Configuration management
# ✅ Production-ready code structure
# ✅ Comprehensive error handling
# ✅ GPU acceleration support
# ✅ Cross-platform compatibility